+++ src/reproducer.rs
@@ -32,7 +32,7 @@ pub fn reproduce(input_file: &Path) -> Result<()> {
     )?;
     let inputs = OpenApiInput::from_file(input_file)?;
 
-    let (authentication, cookie_store, client) = build_http_client()?;
+    let (mut authentication, cookie_store, client) = build_http_client(&api)?;
 
     println!(
         "Input file {:?} contains {} inputs",
@@ -54,18 +54,27 @@ pub fn reproduce(input_file: &Path) -> Result<()> {
             continue;
         };
 
-        let request_built = match build_request_from_input(&client, &cookie_store, &api, &request)
-            .map(|builder| builder.build())
+        let request_built = match build_request_from_input(
+            &client,
+            &mut authentication,
+            &cookie_store,
+            &api,
+            &request,
+        )
+        .map(|builder| builder.build())
         {
-            None => {
-                warn!("Could not generate a HTTP request from this input. Skipping ...");
+            Err(message) => {
+                warn!(
+                    "Could not generate a HTTP request from this input: {}. Skipping ...",
+                    message
+                );
                 continue;
             }
-            Some(Err(message)) => {
+            Ok(Err(message)) => {
                 error!("Error building the request: {}", message);
                 break;
             }
-            Some(Ok(request)) => {
+            Ok(Ok(request)) => {
                 info!(
                     "Converted to CURL command:\n{}",
                     CurlRequest(&request, &authentication)