+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Adds custom executor enabling proper timeout kill and removing unsafe code in
   [#49](https://github.com/TNO-S3/WuppieFuzz/pull/49)
+- Implement support for access and refresh tokens for OAuth in
+  [#70](https://github.com/TNO-S3/WuppieFuzz/pull/70)
 
 ## Fixes